Se arreglaron los bugs de los select, se agrego un campo a puestos y se hicieron modificaciones a nivel de vista

+++ database/migrations/2020_08_23_183427_create_empleados_table.php
@@ -31,9 +31,9 @@ class CreateEmpleadosTable extends Migration
             $table->string('nombre', 35);
             $table->string('apellido', 35);
             $table->date('fecha_nacimiento');
-            $table->integer('dpi');
+            $table->bigInteger('dpi');
             $table->string('direccion', 70);
-            $table->integer('telefono');
+            $table->bigInteger('telefono');
             $table->boolean('estado')->default(1);
 
             $table->timestamps();
